timetable/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1

sqlalchemy==2.0.32
pymysql==1.1.1

requests==2.32.3

ddddocr==1.5.6

pycryptodome==3.20.0

# optional but recommended for complex captcha
opencv-python-headless==4.9.0.80
